Temporarily handle "IOError: invalid data stream"
Seems like urllib2 is not doing the best of the jobs when downloading
data over a very slow network.
I am not very sure as to what would be causing an "invalid data stream"
corruption of the data.
Is it something wrong at my end ? Or something wrong that Python is
doing? Or maybe something just plain wrong with the network ?

Needs investigation.