Explicitly import httplib module
So, in theory, this should have worked, because httplib is imported by
urllib2, which is what we use.

But given the issues that HTTP handling in python has led to, we can run
into the BadStatusLine exception, which is not handled well by Python. I
believe there's already a bug report filed against this oddity.

For apt-offline, for now, let's just keep all the guys happy.

Exception in thread Thread-5:
Traceback (most recent call last):
  File "/opt/python/2.7.9/lib/python2.7/threading.py", line 810, in __bootstrap_inner
    self.run()
  File "/opt/python/2.7.9/lib/python2.7/threading.py", line 763, in run
    self.__target(*self.__args, **self.__kwargs)
  File "/home/travis/build/rickysarraf/apt-offline/apt_offline_core/AptOfflineLib.py", line 636, in run
    self.responseQueue.put( self.WorkerFunction( item, thread_name ) )
  File "/home/travis/build/rickysarraf/apt-offline/apt_offline_core/AptOfflineCoreLib.py", line 848, in DataFetcher
    if DownloadPackages(NewUrl) is True:
  File "/home/travis/build/rickysarraf/apt-offline/apt_offline_core/AptOfflineCoreLib.py", line 812, in DownloadPackages
    if FetcherInstance.download_from_web(url, pkgFile, Str_DownloadDir) == True:
  File "/home/travis/build/rickysarraf/apt-offline/apt_offline_core/AptOfflineCoreLib.py", line 312, in download_from_web
    except httplib.BadStatusLine:
NameError: global name 'httplib' is not defined