Initial list build for checksums
Check for header explicitly at the beginning

Add documentation on how it compares

Use a list instead

Fix synatx error

Large commit. Pay attention to it. Build a checksumList based on the content of the Release file. And when downloading a apt update file, retain its extension type

Whitelist the Release and Signature files

Workin implementation where apt update files and the Release files are
checked for integrity and only then proceeded

Both, a direcotry and a zip file, now work for checksum integrity
verification