Check if src and dest are the same file
This has become a PITA. The comments show I fixed it earlier but the
code does not reflect it. Ultimately, it was reading and writing the
same file, which caused it to go NULL bytes.  :-(

The best fix would have been to open the file in an exclusive mode,
which would have helped avoid all these good-for-nothing problems.

Closes: https://github.com/rickysarraf/apt-offline/issues/90