Set apt package install path to /var/cache/apt/archives
Because irrespective of whether you put the file in archives/ or
partial/,  the file's integrity is checked.

On the other hand, if we put into partials/, apt reports that it has to
download that package even though it doesn't really download because the
file is present in partials/

Another drawback is that becase the file is in partials/, before apt
realizes that it doesn't need to download the file, it still established
a connection to the remote repository. This breaks the offline use
case/workflow