Introduce option --strict-deb-check
With this opiton, which defaults to False, users can mandate the level
of .deb package checksum validation.

Validation of the .deb package is delegated completely to the apt
subsystem. Installation of the .deb pacakge, by bypassing apt, is highly
discouraged and unsupported.

Validation occurs in the following manned for .debs:

If --strict-deb-check is enabled, .debs are copied to partial/
apt, when invoked later, will prompt user about full download of payload
but when user continues further, apt will realize that all payload is
already available locally and nothing needs to be downloaded. It will
then invoke the chekcsum validation routine to validate the .debs that
were made available by apt-offline

If --strict-deb-check is not enabled, which is the default behavior,
.debs are copied to the final apt storage location in archives/
apt, when invoked later, will promprt user that no further download is required.
An exception about further download will occur only if the .deb package is tampered.
Note: apt, in this case where the payload is copied to archives/, does validation
check only for the file size of the .deb file.

See Debian Bug for details:
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=871656#30