Give better exit code
When items in threads fail, they could fail for multiple error types,
like: 404, 504, Socket Timeout, etc.

But overall, for apt-offline's operation, we retry with other items.

As an overall operation, we want to continue. But we need to report the
user that there are failures. Also, our exit code needs to be non-zero
to emphasize the same in scripts.

See discussion in: https://github.com/rickysarraf/apt-offline/issues/103