Drop redundant import of apt module
We prefer to use the python apt module but it is not a hard dependency.
Thus, we check if the module is available. And if not, we fallback to
non-python mechanism.

So, from a Debian point of view, the python apt module is a good
candidate for Recommends, which is being set with this upload

Closes: #959085