Drop the unused IgnoredBugTypes
This was never in use.

And if we are to really use fine grained bug downloads, it is going to
be not tidy. Because there are mulitple combinations of bug attributes
that can be used; like Severity, Status etc.

These combinations will further complicate with not real advantage. So
it is better to simply download bug reports per package name