Make the copying of changelog lines less brittle
Stop copying changelog lines when a changelog header containing the exact
version is seen. Until now, random other versions that happened to contain
the installed version did this. Also, any non-header lines that happened to
mention the installed version also did this.

This isn't the most correct fix though, the right way to do this would be
to reimplement the code using proper changelog parsing from the Python
debian module.