Explicity set to format type string
Because from the GUI the input comes are textInput and then later we
validate that the input is integer. But the argument, after validation,
is passed as a string to apt_offline_core.

OTOH, in argparse, we do take the input as an integer. Thus the int decimal
format specified. But, as in this bug report, when using the GUI, it'd
result in a TypeError.

So better to, for display sake, explicitly convert to string.

Closes: https://github.com/rickysarraf/apt-offline/issues/129