Check for the generated signautre file
In cases where there is no payload to download, because:
a. The payload is already installed
b. The payload is already in apt's cache

apt will give a clean exit status. But for apt-offline, the generated
signautre file is 0 bytes.

Check for such case and report an error to the user

Closes: https://github.com/rickysarraf/apt-offline/issues/141