Don't mandate a default option
If a user doesn't provide any option, exit with an error.

The rationale for this change is that the user must specify an option to
the set command, which could be either of update, upgrade,
install-packages etc.

We do this change, to explicitly ask the user for an option input,
because the previous behavior of running update and upgrade together as
the default action, causes too many confustions to the user.

Because the generated signature will always fetch new apt indices
payload, which in turn will have the potential to always make new
upgrade available.

This breaks the single trip workflow, resulting in some download payload
pending.

For some context, see:

https://github.com/rickysarraf/apt-offline/issues/136