Don't append errors on apt update indices metadata
The reason for not appending these errors to the errlist is because apt
generates the signature for multiple file types, depending on your
system. Like:

* Translations
* Components
* Contents
* icons

These are dependency on various tools a user may/not install on their
systems. Then, there's also the remote repository, which may/not host
those files.

These files aren't the important ones for apt to function for
update/upgrade operation.

The only imporntant files for apt update/upgrade operation are:

* Packages
* InRelease
* Release

So, that is why we will not error out on these failure types. Instead,
we'll just log them in verbose mode