Treat 504 as permanent errors
504 errors are special as in they are gateway timeouts/unreachable.

When a proxy is unreachable, it reports with a 504. As such, we need to
consider such errors and report them.

On the other hand, the reason we are treating 404 as a retriable error
is because of apt update indices, where we'll get a 404 but still want
to retry with other supported compression formats

As for downloads like .deb files, they are handled confitionally when
the url endswith '.deb'. And the error is reported immediately.

Without this change, all my proxy errors were unreported because they
were being treated as retriable. Certainly not the desired output one
would expect.