Call stripper() under a try/except block
Because the signature file is generated by apt. And depending of the
distribution, it could have some extra changes added in it. Like in this
linked github issue report.

Nevertheless, given that the apt output is not in a declared machine
readable format, nor is it officially recommended to parse the output;
instead let us make apt-offline put more checks in place when parsing
its input.

Closes: https://github.com/rickysarraf/apt-offline/issues/199