check for checksum and handle appropriately
Note that checksum is not always available in the signature file. For
example, for the `--update` operation, there's no checksum in the
signature. Instead, that is handled separately via GPG Repository
Singing.

The checksums are only present for the build package files, like the
.deb files. That's because in Debian, we only sign the repository
metadata and list all checksums and other important bits, as part of
that metadata

Closes: https://github.com/rickysarraf/apt-offline/issues/216
Thanks: Dan Whitman (Github:kyp44)