Retry when server reports 429 Too Many Requests occurs
An HTTP/1.1 429 Too Many Requests occurs when:

* there are too many requests
* server is overloaded

The server may respond back with a `Retry-after` field with the value in
seconds. For now, we'll just retry immediately. A potential enhancement
could be to honor `Retry-after` but, thinking of it, it doesn't fit the
bill for `apt-offline` because we want the download operation to
conclude immediately rather than wait for `Retry-after` seconds.

Thanks: Zoltan Kelemen (Github: misterzed88)
Closes: https://github.com/rickysarraf/apt-offline/issues/213